Fix from "Refactoring ACC CUD parts."

+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/service/ExtensionService.java
@@ -65,7 +65,7 @@ public class ExtensionService {
     public CcAccNode getExtensionNode(User user, long manifestId) {
         AccManifestRecord extensionAcc = getExtensionAcc(manifestId);
 
-        CcAccNode ueAcc = repository.getAccNodeByAccManifestId(user, extensionAcc.getAccManifestId());
+        CcAccNode ueAcc = repository.getAccNodeByAccManifestId(user, extensionAcc.getAccManifestId().toBigInteger());
         CcAsccpNode asccpNode = repository.getAsccpNodeByRoleOfAccId(ueAcc.getAccId(), extensionAcc.getReleaseId());
         if (asccpNode == null) {
             return null;
